requirements.txt
openai
PyMuPDF
faiss-cpu
numpy
python-dotenv
langchain
langchain-openai
langchain-community
tqdm
